requirements.txt
discord.py
requests
google-api-python-client
beautifulsoup4